backend/requirements.txt
fastapi==0.101.1
pydantic==2.2.1
httpx==0.28.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==1.68.2
